Removing old variables implementation and moving variables_new to variables
